move partial tests
